requirements.txt
langchain
langchain-openai
langchain-chroma
chromadb
python-dotenv
